refactoring : 모달 열리고 닫히는 트리거 함수 생성
