+++ src/features/chat/ui/ChatRoomList.tsx
@@ -5,6 +5,8 @@ import ChatRoomItem from './ChatRoomItem';
 import useSubChatRooms from '../hooks/socket/useSubChatRooms';
 import ChatSearchBar from './ChatSearchBar';
 import { useEffect, useState } from 'react';
+import Image from 'next/image';
+import useChatDialogTrigger from '../hooks/useChatDialogTrigger';
 
 interface IChatRoomListProps {
   selectedRoomId: string;
@@ -14,6 +16,7 @@ export default function ChatRoomList({ selectedRoomId }: IChatRoomListProps) {
   const { rooms } = useChatWebSocketStore();
   const [searchQuery, setSearchQuery] = useState('');
   const [filteredRoom, setFilteredRoom] = useState(rooms);
+  const { closeChatDialog } = useChatDialogTrigger();
 
   const handleSearchRoom = (value: string) => {
     setSearchQuery(value);
@@ -27,6 +30,15 @@ export default function ChatRoomList({ selectedRoomId }: IChatRoomListProps) {
 
   return (
     <section className="h-full w-full flex-1/5 border-r border-border_primary/20">
+      <div className="flex items-center justify-between p-4 border-b border-border_primary/20">
+        <h2 className="text-xl font-semibold text-white">채팅</h2>
+        <button
+          onClick={closeChatDialog}
+          className="p-2 hover:bg-white/10 rounded-[10px] transition-colors"
+        >
+          <Image src="/icons/close/closeWithBorder.svg" height={20} width={20} alt="closeBtn" />
+        </button>
+      </div>
       <ChatSearchBar searchQuery={searchQuery} onSearch={handleSearchRoom} />
       {rooms.length > 0 ? (
         <ul>
